caps/test: address review — no magic numbers, no algorithm-specifics (#45)
Review follow-up on PR #57:

- Derive the advertisement cap from the master-list size instead of a magic 64:
  HYBRID_MAX_TLS_GROUPS = HYBRID_KEM_ALG_COUNT,
  HYBRID_MAX_TLS_SIGALGS = HYBRID_SIG_ALG_COUNT. The invariant is "never
  advertise more distinct entries than are defined"; it tracks the lists.

- Classify code-point provenance by VALUE, not by algorithm name. Add
  hybrid_codepoint_is_iana_assigned() for the draft-ietf-tls-ecdhe-mlkem span
  (0x11EB-0x11ED) and drop the hardcoded MLX name/code-point table from the test;
  every entry must now be IANA-assigned-span or provisional-range by value alone.

- Make the operability test algorithm-agnostic: instead of naming Frodo/BIKE/
  Falcon/ML-KEM/ML-DSA, iterate the master tables and cross-check the provider's
  advertisement against an independent component-fetchability probe, in both a
  default-only and a default+oqsprovider context. No specific algorithm is
  assumed present.

- Replace magic name-buffer sizes (80/64) with a shared HYBRID_ALG_NAME_MAX and
  prove sufficiency: the collectors flag truncation and the test asserts none
  occurred. Size the tls_test capture array from HYBRID_KEM_ALG_COUNT. Rename the
  collector capacity to a documented, overflow-guarded MAX_ADV.

- The hybrid-sigalg bound counts only hybrid-table entries (the composite family
  advertises into the same enumeration but is bounded by its own list).

- Explain why table_sig_cp is #ifdef'd (unused-function on OpenSSL < 3.2).

+++ hybrid_caps.c
@@ -269,7 +269,7 @@ int hybrid_get_capabilities(void *provctx, const char *capability,
             }
             if (advertised >= HYBRID_MAX_TLS_GROUPS) {
                 hybrid_log("hybrid: TLS-GROUP advertisement cap (%d) reached; "
-                           "dropping %s (0x%04x)\n", HYBRID_MAX_TLS_GROUPS,
+                           "dropping %s (0x%04x)\n", (int)HYBRID_MAX_TLS_GROUPS,
                            name, hybrid_group_list[i].group_id);
                 continue;
             }
@@ -300,7 +300,7 @@ int hybrid_get_capabilities(void *provctx, const char *capability,
             }
             if (advertised >= HYBRID_MAX_TLS_SIGALGS) {
                 hybrid_log("hybrid: TLS-SIGALG advertisement cap (%d) reached; "
-                           "dropping %s (0x%04x)\n", HYBRID_MAX_TLS_SIGALGS,
+                           "dropping %s (0x%04x)\n", (int)HYBRID_MAX_TLS_SIGALGS,
                            name, hybrid_sigalg_list[i].code_point);
                 continue;
             }
